requirements.txt
flask==3.0.2
flask-socketio==5.3.6
eventlet==0.35.2

paho-mqtt==1.6.1
python-dotenv==1.0.1

SQLAlchemy==2.0.27
mysql-connector-python==8.0.33
